[Picture in Picture] Add event handler for custom controls click
This change adds onpictureinpicturecontrolsclick to the partial
interface of an HTMLVideoElement and creates the plumbing that
allows for functions to trigger that event (note: it is not
currently being called in this CL).

Manual testing was done by initiating the call in
overlay_window_views.cc by calling controller_->ControlClick() and
confirming the new EventHandler was triggered.

Bug: 852572
Change-Id: I57038fa4809085b7000ca7e5d933c38e82d133e7
Reviewed-on: https://chromium-review.googlesource.com/1100043
Commit-Queue: Steven Sawtelle <sawtelle@google.com>
Reviewed-by: apacible <apacible@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: François Beaufort <beaufort.francois@gmail.com>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#570320}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 42eac6a83fc486c1263a7b2c7858312d08c6fb09