Improve buffer size checking in MojoSharedBufferVideoFrame
If the U and V data is interleaved in the buffer, then the last row of data
doesn't require a full stride. So only count RowBytes() for the last row
when computing where the end of the data is.

BUG=829443
TEST=new media_mojo_unittest passes

Change-Id: I186e4dd9d2045f34897bbf2c1d71e04421e199b7
Reviewed-on: https://chromium-review.googlesource.com/1114335
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#571001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95907d880e2bf2264fba79d06d3bb56576329748