[Image Capture, Android] Fix for "UnknownError: platform error"
Fixes an issue where takePhoto would occasionally fail due to a Java object
being released too early.

In particular, an ImageReader instance is created and its Surface object is
held on to and used asynchronously while the ImageReader goes out of scope and
may get garbage collected. The fix is to hold on to the ImageReader instance
for as long as the Surface may be used.

Bug: 846576
Change-Id: Icb0e3b247a1a75af8ce86c44774960c2b036083b
Reviewed-on: https://chromium-review.googlesource.com/1117854
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#571574}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 356a67ed5facf20d6a52748444bb580fe73dc329