[Image Capture] Add stress test that takes 10 photos in a row
The purpose of the added test is to ensure the reliability of the feature as
well as detecting flakiness that could be caused by race conditions.

The test is marked as MANUAL_ to avoid it running on the regular commit queue
trybots.

To run the test manually on Android, increasing various timeouts is needed:
out/android_release/bin/run_content_browsertests --gtest_filter=*WebRtcImageCaptureStressBrowserTest* --test-launcher-retry-limit=0 --ui-test-action-max-timeout=60000 --shard-timeout=60
The above sets the timeout to 1 minute, which should be more than enough for 10 photos.

Bug: 857530
Change-Id: Ibf4f78d43dc3cbe24687710f48bf311f7bb0bfd4
Reviewed-on: https://chromium-review.googlesource.com/1135703
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#575697}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5240a33ded02b4e0b90b235dae805e07f5cd39a4