[Video Capture] Move static factory logic out of class VideoCaptureDeviceFactory
Moves the existing methods for creating a platform-specific instance of interface
media::VideoCaptureDeviceFactory out of the interface class and into standalone
functions in a separate file.

This is reasonable, because the factory logic does really not belong in the
abstract interface.

This is needed to be able to move the factory logic part and the
VideoCaptureDeviceFactory interface to separate gn targets in the subsequent
CL.

This CL is part of a series, see Design Doc at
https://docs.google.com/document/d/1mYnsZfLBRmbsDpUtfb6C7dzhfw2Kcxg_-uiG_6MnWVQ/edit?usp=sharing

Change-Id: Iddd093c3a8d0b665094ca3ded32ff4c46f592219
Reviewed-on: https://chromium-review.googlesource.com/1139175
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#576497}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e1bfc30e0e29c128215130e2729ff0dc4384be16