[Video Capture, Android] Catch AssertionError on getCameraCharacteristics
Crash reports indicate that on some devices the call to getCameraCharacteristics
raises an AssertionError inside Android OS level code. The message attached to
it is typically "Supported FPS ranges cannot be null."

On the Chromium side, the best we can do is to catch this, so that instead of
crashing, we fail gracefully. This CL adds the catch clause.

Bug: 860784
Change-Id: Ic5092e98b9497399ac309f00ea3eaecd41512646
Reviewed-on: https://chromium-review.googlesource.com/1145727
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#577224}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bcde09d4a85890dac489b8cb66ae762204039682