limit parallel media preloading
When loading many media files, we normally end up having to wait
for the first request for ALL media files to load before we can
process the second request of the first media file. By limiting
the number of parallel preloads, time-to-first frame should improve
for all but the last media file.

However, this feature is checked in behind a feature flag, so that
we can run some expriments on it before turning it on by default.

Bug: 884820
Change-Id: Ib0c366325d82fb59e6ad896bf985b1e828143a82
Reviewed-on: https://chromium-review.googlesource.com/1228486
Commit-Queue: Fredrik Hubinette <hubbe@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#594090}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 395ca8adef07ee3e2a8644dbb9c5725373f246b2