media: Fix use of |weak_this_| in MojoCdmService
This was added accidentally in the original CDM, not initialized, and
not used. Then in a late CL it's used assuming it's initialized, which
was wrong.

This CL removes |weak_this_| from MojoCdmService to fix the issue.

Bug: 432998
Change-Id: Ifcda68e79c7ca5f2203cad9e77765b7f8c3d1971
Reviewed-on: https://chromium-review.googlesource.com/1255464
Reviewed-by: John Rummell <jrummell@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#595568}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 92f9e46505f4cfe384c22c168f141cd4e79fbb1b