media: Convert media/test/data/README to a markdown doc
This makes it easy to organize the doc and find files.

This CL does minimal reordering of the files listed in the current
README, with some rough grouping. It is expected that we'll
improve the structure of the file gradually over time.

Bug: 818048
Change-Id: I8f19fa388d6cd599fc32824b91d184bd49b2767a
Reviewed-on: https://chromium-review.googlesource.com/1253207
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#595609}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f0257d49eae32a43d024b63855eb57db4c025caf