Disable "pause on hide" when flinging videos
Currently, we sometimes pause videos when they are hidden, as an
optimization. This causes videos flung to cast devices to pause as well,
since the pause command sent to the FlingingRenderer is forwarded to the
cast device.

This CL fixes adds a new flag to track whether or not we are currently
flinging a video, and disables ShouldPauseVideoWhenHidden() when we are.

Bug: 790766
Change-Id: I7729ab2c8187ffd1c77ebb6c48301165dea1d90f
Reviewed-on: https://chromium-review.googlesource.com/c/1292820
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#601386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29ae1a90775c019ecba7d7386c28beaa8b521f32