Use SurfaceLayer in kOnDemand mode if player starts in PIP.
When the SurfaceLayerMode is kOndemand, WMPI would wait until
entering PIP before switching to SurfaceLayer.  However, it
missed the case where it started in PIP mode from a previous player.

This CL adds a check to see if the client is already in PIP mode,
and switches to SurfaceLayer immediately if so.

Bug: 891619
Change-Id: If1d75982ebcfb3c633814fd8247470a8a595c3ef
Reviewed-on: https://chromium-review.googlesource.com/c/1300093
Reviewed-by: Fredrik Hubinette <hubbe@chromium.org>
Commit-Queue: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#602883}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 702d721db7ac886392fb32774292c56cfe8d10fe